get_service_target_report

+++ Reports/serializers.py
@@ -960,7 +960,7 @@ class ServiceGroupReport(serializers.ModelSerializer):
             for ord  in service_target:
                 created_date = ord.year.date() 
                 if created_date.month == date_obj.month and created_date.year == date_obj.year:
-                    ser_target += int(ord.service_target)            
+                    ser_target += float(ord.service_target)            
             return ser_target
             
         except Exception as err:
